snis_client: make planet shade a smooth float instead of an int flag
The analytic planet umbra and penumbra shading computes a fractional
occlusion, but the flat shaded single-color-lit-per-pixel program declared
u_in_shade as an int.  glUniform1f() uploads a float, which the driver
truncates to int, so every penumbra fraction between 0 and 1 collapsed to
0, fully lit, and only a fraction of exactly 1.0, deep umbra, read as
shaded.  The result was a hard pop from fully lit to fully shadowed at the
umbra rather than a gradual fade through the penumbra.  The old binary
shading only ever sent 0.1 or 1.0, which truncated correctly, so the int
type had gone unnoticed, and the textured-and-lit program already used a
float and faded smoothly.

Declare u_in_shade as a float in both the vertex and fragment stages of the
per pixel program (matching across the linked pair is required), for
desktop and GLES, and drop the now unnecessary float() cast.  Change
object_in_shade() to return the raw fraction, 0.0 lit to 1.0 umbra, instead
of the 0.1 baseline that existed only so the value truncated to 0.  Flat
shaded objects are unchanged at the lit and umbra endpoints and now fade
correctly through the penumbra; the ambient term still floors the umbra.
Textured lit objects, which previously sat at the 0.1 baseline (about 90%
diffuse) when fully lit, now reach full brightness, removing a pre-existing
flat versus textured inconsistency.

+++ share/snis/shader-es/single-color-lit-per-pixel.frag
@@ -1,6 +1,6 @@
 
 uniform vec3 u_LightPos;       // The position of the light in eye space.
-uniform int u_in_shade;        // 0 means not full shade, 1 means full shade
+uniform float u_in_shade;      // 0.0 means not in shade, 1.0 means full shade (penumbra in between)
 
 varying vec3 v_Position;       // Interpolated position for this fragment.
 varying vec3 v_Color;          // This is the color from the vertex shader interpolated across the triangle per fragment
@@ -17,7 +17,7 @@ void main()
 	float dot = dot(v_Normal, lightVector);
 
 	// mimic the original snis software render lighting
-	dot = (1.0 - float(u_in_shade)) * ((dot + 1.0) / 2.0);
+	dot = (1.0 - u_in_shade) * ((dot + 1.0) / 2.0);
 
 	// ambient light
 	float diffuse = max(dot, u_Ambient);
